+++ __tests__/route-match.test.tsx
@@ -20,7 +20,7 @@ import {
     useParams,
     Outlet,
     useRoutes,
-} from '..'
+} from '@'
 
 describe('route matching', () => {
     function VueFundamentals() {