+++ __tests__/router-dom/search-params.test.tsx
@@ -8,7 +8,7 @@ import {
     useNavigate,
     useParams,
     useSearchParams,
-} from '../..'
+} from '@'
 
 describe('useSearchParams', () => {
     const SearchPage = defineComponent({