+++ lib-vs-rrv5/index.ts
@@ -1,6 +1,0 @@
-import MemoryRouter from './MemoryRouter'
-import Route from './Route'
-import Router from './Router'
-import Switch from './Switch'
-
-export { Router, Route, Switch, MemoryRouter }